+++ UnityUtils-UnityProject/Packages/UnityUtils/Runtime/Script Utils/Objects/ModifierType.cs
@@ -3,11 +3,15 @@
   /// Enum with different modifier types that get used in <see cref="ObjectModifiers{T}"/> when applying modifiers to an input value.
   /// </summary>
   public enum ModifierType {
-    /// <summary> Addition/Subtraction modifier </summary>
+    /// Addition/Subtraction modifier
     Flat,
-    /// <summary> Multiplication modifier </summary>
+    /// Multiplication modifier
     Multiply,
-    /// <summary> Division modifier </summary>
-    Divide
+    /// Division modifier, divides the input value by the modifier value. If value is 0, it will be ignored to avoid dividing by zero errors.
+    Divide,
+    /// Root Modifier, gets the root of the input value based on the modifier value. 
+    Root,
+    /// Exponent Modifier, raises the input value to the power of the modifier value.
+    Exponent
   }
 }